refactor: Move firebase.js to scrpts folder
