+++ public/scripts/classes/firebase.js
